feat: derive key and register worker
