+++ package.json
@@ -22,6 +22,7 @@
     "dotenv": "^16.4.7",
     "joi": "^17.13.3",
     "near-api-js": "^4.0.3",
+    "near-seed-phrase": "^0.2.1",
     "node-cache": "^5.1.2",
     "winston": "^3.3.3",
     "ws": "^8.14.2"
